move views.index and views.stats to views.common
